Leave CallVisualizer class in 'core' package until major release - Rename addVisitorContext() function parameter so that it reflected its purpose - Minor improvements in MainFragment
MOB-4360
