+++ widgetssdk/src/main/java/com/glia/widgets/GliaWidgets.kt
@@ -10,7 +10,7 @@ import com.glia.widgets.callbacks.OnError
 import com.glia.widgets.callbacks.OnResult
 import com.glia.widgets.chat.adapter.CustomCardAdapter
 import com.glia.widgets.chat.adapter.WebViewCardAdapter
-import com.glia.widgets.callvisualizer.CallVisualizer
+import com.glia.widgets.core.callvisualizer.domain.CallVisualizer
 import com.glia.widgets.liveobservation.LiveObservation
 import com.glia.widgets.queue.Queue
 import com.glia.widgets.queue.toWidgetsType
