+++ widgetssdk/src/main/java/com/glia/widgets/internal/callvisualizer/CallVisualizerManager.kt
@@ -4,7 +4,7 @@ import android.annotation.SuppressLint
 import android.content.Context
 import android.view.View
 import com.glia.widgets.callvisualizer.controller.CallVisualizerContract
-import com.glia.widgets.callvisualizer.CallVisualizer
+import com.glia.widgets.core.callvisualizer.domain.CallVisualizer
 import com.glia.widgets.internal.callvisualizer.domain.VisitorCodeViewBuilderUseCase
 import com.glia.widgets.helper.unSafeSubscribe
